Add task polling to starving workers
In narrow regions of the dependency graph, there are too few tasks to
saturate the worker pool and as a result, workers (unavoidably) starve.

Previously the starving worker would shutdown, making it unavailable in
the event that the dependency graph opens up again. This commit makes
the workers to continuously poll the queue (with a small timeout) so
that the resources will be available when needed.